+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## Unreleased
+
+### Added
+
+-- `code42 legal-hold events` command:
+    - `matter-id` filters based on a legal hold uid.
+    - `begin` filters based on a beginning timestamp.
+    - `end` filters based on an end timestamp.
+    - `event-type` filters based on a list of event types.
+
 ## 1.4.0 - 2021-03-09
 
 ### Added